fix sudo behavior detection
The purpose of can_passwordless_sudo is to decide if a password is
needed to use sudo.

As it was used, ``sudo -n`` command determined that no password is
needed if sudo is previously used in the same terminal and the sudo
timeout is not expired.

Using ``-k`` flag allows to ignore any registered sudo token. It does not
delete any existing token (so it preserves sudo "state"). It seems a
better choice without any regression or behavior change.