fix on previous PR #59
fixed code is duplicated. Previous PR only fix one occurence, and
probably the wrong one as I find no occurence of env_creator (previously
fixed file) in nsbl or freckles.

I surely wrongly reported fix from my working env to my git repository.

I test this fix, and it ensures the following and correct behavior:

* if passwordless sudo, no prompt for sudo password
* if password sudo, prompt for password even if a sudo token is
  available and not expired, so that long freckles run do not fail on
  token timeout.

Before this fix, password is not prompted if a token is available and
not expired.